code/basecoding-genai-bento/requirements.txt

--extra-index-url http://proget.spins.team/pypi/private-pypi/simple
--trusted-host proget.spins.team

annotated-types==0.6.0
    # via pydantic
anyio==4.2.0
    # via
    #   httpx
    #   openai
bentoml>=1.0.0

cachetools==5.3.2
    # via google-auth
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
distro==1.9.0
    # via openai
fuzzywuzzy==0.18.0
    # via -r requirements/base.in
google-api-core[grpc]==2.17.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.27.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-bigquery==3.12.0
    # via -r requirements/base.in
google-cloud-core==2.4.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-secret-manager==2.18.2
google-cloud-storage==2.11.0
    # via -r requirements/base.in
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.7.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.62.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.60.1
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   grpcio-status
grpcio-status==1.60.1
    # via google-api-core
h11==0.14.0
    # via httpcore
httpcore==1.0.2
    # via httpx
httpx==0.26.0
    # via openai
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
levenshtein==0.24.0
    # via python-levenshtein
numpy==1.26.0
    # via
    #   -r requirements/base.in
    #   pandas
openai==1.11.1
    # via -r requirements/base.in
packaging==23.2
    # via google-cloud-bigquery
pandas==1.5.3
    # via -r requirements/base.in
proto-plus==1.23.0
    # via google-cloud-bigquery
protobuf==4.25.2
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pydantic==2.5.1
    # via
    #   -r requirements/base.in
    #   openai
pydantic-core==2.14.3
    # via pydantic
python-dateutil==2.8.2
    # via
    #   google-cloud-bigquery
    #   pandas
python-dotenv==1.0.0
    # via -r requirements/base.in
python-levenshtein==0.24.0
    # via -r requirements/base.in
pytz==2024.1
    # via pandas
rapidfuzz==3.6.1
    # via levenshtein
regex==2023.12.25
    # via tiktoken
requests==2.31.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   tiktoken
rsa==4.9
    # via google-auth
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
    #   openai
tiktoken==0.5.2
    # via -r requirements/base.in
tqdm==4.66.1
    # via openai
typing-extensions==4.9.0
    # via
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.2.0
    # via requests
